requirements.txt
# Core Python dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM and AI
openai>=1.6.1
langchain==0.0.350
langchain-openai==0.0.2
tiktoken==0.5.2

# Data processing
pandas==2.1.4
numpy==1.24.3
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2

# Resume parsing (compatible version)
resume-parser==0.8.4

# Web interface
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0

# Document processing and formatting
reportlab==4.0.7
jinja2==3.1.2
markdown==3.5.1

# Database (for enhanced features)
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-multipart==0.0.6

# Optional: Enhanced features
# redis==5.0.1
# elasticsearch==8.11.0
# scikit-learn==1.3.2
# transformers==4.35.2 